Moving ./src/phSim.js to ./src/index.js.
